[FIX] l10n_es_aeat_mod349: Use first period year in export
Relying in fiscal year code for getting the fiscal year is not safe,
because you can use that code for defining your journal sequences (for
example, with the module account_auto_fy_sequence).

We can't use also the year of the fiscal year, because you can have
split fiscal years (i. e. like the academic courses). What we uses then
is the year of the first period of the declaration, as there's no
legal possibility of having mixed periods from different years.