[FIX] l10n_es_aeat_sii: Constraint not suitable for multi-company
This constraint leads to integrity errors due to missing values in companies that
doesn't have SII enabled. As the required attribute is set in view, this should be removed.

There's still the chance of writing an invoice without this key by code, but putting
a constraint seems excessive as this key is only needed when sending the invoice. If
not present, there will be an SII error that will help you to identify the problem.