[FIX] l10n_es_aeat_sii: Avoid access error to queue.job records
Putting groups in the view is not enough for preventing Odoo to load the linked one2many model,
resulting in an error for those without permissions in queue.job object. The solution for this
is to isolate that changes in a separate view with a group, and this way the fields are not
fetched anymore if the user doesn't belong to that group.

There's other view group that also has the same group (AEAT responsible), but shouldn't be
isolated, as there's a field there (sii_state), that it's needed for the visibility of
other elements, and there's no problem with them, as they are not x2x fields.