[ADD] l10n_es_account_group: Spanish Account Groups
This includes account groups for Spanish chart of accounts.

Installation
============

If you have already the chart of accounts loaded on your company, you will
need to update it through the module `account_chart_update`. Follow the
instructions on that module for that.

Usage
=====

For reviewing groups assigned to accounts:

* Go to *Invoicing > Adviser > Chart of Accounts*.
* Open one account and check "Group" field.

Known issues / Roadmap
======================

* This module shouldn't be migrated to v11, as it's a native feature.