[FIX] l10n_es_aeat: Create sequences when new company is created
When a new company is created, its sequences are not populated, so an Odoo service
restart is needed for being able to use AEAT models. With this, you don't need
this restart.

It includes a test with a hack for populating a model in the test itself for testing
complete flow.