[FIX] l10n_es_account_bank_statement_import_n43: Change encoding for compatibility with BBVA files (#1056)
Some statement files include BOM data (https://es.wikipedia.org/wiki/Marca_de_orden_de_bytes) at the beginning of the file, making them non readable by current parser, as fixed positions are moved in these cases and thus expected data is not matched.

As ASCII, which is the encoding used in files according specification, is a subset of UTF-8, there's no danger in using it with the variant 'utf-8-sig' for handling if BOM marker is present or not.