[FIX] l10n_es_account_bank_statement_import_n43: Don't fail on other formats
If other statement format is present, the file is properly decoded and stored
in `data_file` variable, but if the format is not N43, the loop will continue
and will try to decode the already decoded content in other encoding, giving an
error. We avoid that storing in a different variable the decode content.