[FIX] l10n_es_facturae: Handle other price unit digits different from 2
It turns out the information in the XML file gets incoherent in this case, having
`GrossAmount` and `TotalCost` fields with different rounding.

After several attempts, the best approach is to simply apply the same rounding in the
`TotalCost` (subtotal before discount) field (got from currency), and compute the
discount amount substracting both.

This can lead to (`TotalCost` * <discount percentage> / 100) != <discount amount>, but
there's no other way to make the rest correct, and services don't seem to perform
this check, but the other, as these fields are mostly informative.