[FIX] l10n_es_aeat_mod303: Include both signs (debit and credit) in tax fee fields
Given this regular invoice:

1, IVA 10%, 10 €, 1 € tax
2, IVA 21%, -1 €, -0,21 € tax

that can happen when you return some packages (like returnable bottles), you have
a valid invoice with total positive amount to due, but in 303, the -0.21 € of the
tax fee is not included as there's no other positive tax fee to compensate that
negative amount, and the tax map line indicated that only credit operations are
included.

As the move type is already checked for avoiding to include refunds, we can change
to include both move lines types (debit and credit) for this. In fact, in the
base, both are already included.

As this can happen as well with other taxes and with supplier invoices, all
involved fields are changed in the same commit.