[FIX] l10n_es_aeat_mod347: Handle VAT correctly
Since v11, Odoo doesn't need the country code in VATs for performing the checks. You
must fill country + putting the VAT with or without the country code.

This model doesn't work properly for the case where you don't introduce the
country code.

We now use the method created for centralizing all the partner identification stuff
that was introduced with VAT book for performing a proper recognition.

Included tests for checking all of the identification stuff.