[FIX] l10n_es_aeat_mod347: Don't remove inactive partners from the report
It seems ORM is removing inactive records when filtering by a submodel domain,
specifically ('commercial_partner_id.not_in_mod347', '=', False), as this expression
is converted in an `IN` SQL clause that pre-filter records, which leads to an
incomplete report although the limits were higher.

We solve it passing the active_test=False context for evaluating all partners.