[FIX] l10n_es_aeat_sii_oca: Avoid KeyError on taxes with no amount
On taxes with no amount ("IVA soportado exento" for example), the key `amount` is not
filled, provoking a problem sending it to SII.

We avoid it initializing full expected dictionary structure on first occurrence.

This commit also contains a little fix for avoiding the crash of `merge_dict` method
not having "TipoImpositivo" key in the dictionaries.