[FIX] l10n_es_aeat_sii_oca: Change key when invoicing purchase order
Steps to reproduce:

1. Create vendor with fiscal position "Régimen intracomunitario".
2. Assign in the fiscal position the SII registration key for purchases
   "[09]".
3. Create a purchase order for that vendor and confirm it.
4. Press on "Create Bill" button.

Current behavior:

The vendor bill is created with SII registration key "[01]".

Expected behavior:

The vendor bill is created with SII registration key "[09]".

This is provoked by the sequence Odoo follows for creating the vendor
bill, using a combination of default values + the triggering of
onchanges. The problem is that the onchange of the fiscal position is
triggered once at the beginning (when no partner nor fiscal position
is set), and not triggered again when this happens.

The solution is to catch on the partner onchange also the possible
condition for changing the SII registration key, and trigger it
manually.

Unfortunately, a regression test can't be added, as this module doesn't
depend on purchase for recreating the steps above.

Fixes ##1524