[FIX] l10n_es_aeat_sii_oca: SII description override
If an invoice has de "sii_description" attribute set to "False", this will override the default_description computed when creating the invoice and produce an error when trying to treat a Bool as a String. We should check if it is defined or not and prevent the possibility of it being set to False.

TT28236