[FIX] l10n_es_aeat_sii_oca: Hook on post low level method
We were intercepting the high level method for posting the invoice,
but this method calls a low level one that is the one that is always
called.

There's a chance of having the high level method to be bypassed and
not called, so we need to intercept the low level one.

We also need to filter the non invoices records now at this level.