[FIX] l10n_es_vat_book: Restrict tax move lines to certain accounts
On intra-community operations and other 2-sided received taxes, we have
these move lines:

| Account  | Tax                            | Debit   | Credit  |
|----------|:-------------------------------|:--------|:--------|
| 47200000 | IVA 21% Adquisición de serv... | 16,63 € | 0,00 €  |
| 47700000 | IVA 21% Adquisición de serv... | 0,00 €  | 16,63 € |

so both lines were nullified when computing tax amount.

We need to restrict the search of this kind of lines to one side, and
the best way to do it is through linked account, in a similar way as
it's done on AEAT reports.