[FIX] l10n_es_aeat_mod390: Consider all moves (regular + refund) on certain fields
There are certain fields that doesn't split refunds and regular amounts,
so we need to cover it in the same field, having refunds that decrease
the amount on the field. The only currently applicable one is the field
230, as it's the only with mapped taxes, but this commit changes
preventively other fields with currently no taxes to the same type in
case anytime are populated.

TT34123