[FIX] l10n_es_aeat_mod347: Preserve balance sign on partner move records
When populating move records for a partner, an abs operation is
performed, as AEAT requires to express amount as positive for all
records.

Previous code trusts on the journal entry type for determining later
the sign of the operation, but there can be the possibility of having
a big journal entry where the type is of one sign, but the specific
operation for that partner is another. The canonical example is on PoS,
with the closing entry including together all the operations of the
session.

So let's keep the balance sign all the time, and remove the checks on
move type, converting the amount to positive where needed (BOE export),
and this also makes a bit lighter the quarter computation, not needing
to filter by move_type.

It includes the migration scripts for adapting following the same
previous approach the existing amounts.

TT34561