[FIX] delivery_seur: label filename
In v12, with the parameter datas_fname we could force the file to be
downloaded as such filename. Now that parameter doesn't exist and the
filename format is lost from the previous version.

odoo/odoo@c212cfe8992292f57acacd809de26e1817533095

This is specilly problematic with txt format and some printer which
doesnt support joliet filenames or even files without extension.

The mimetype for text files was also wrong, as 'application/txt' doesn't
exist. See Mozilla Dev docs for more info on this:

https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/MIME_types/Common_types

TT34599

delivery_seur 13.0.1.1.2