[FIX] payment_redsys: Use updated library + black formatting
pycrypto is unmaintained since several ago, and in fact pycryptodome was
the library being used in previous version. The API we are using is
compatible between both libraries though.

There's no problem for already deployed systems, as the check on the
external dependencies is only done at install.

We took the occasion to remove the try mechanism that is not needed now,
although the linter emits a warning.