[FIX] l10n_es_aeat: Consider correctly French overseas
For French overseas territories
(https://en.wikipedia.org/wiki/Overseas_France), there are no
intracommunity fiscal rules, but the country code that AEAT requires
should be FR, as it can be seen here:

https://sede.agenciatributaria.gob.es/Sede/todas-gestiones/impuestos-tasas/declaraciones-informativas/modelo-347-decla_____racion-anual-operaciones-personas_/instrucciones/codigos-paises-territorios.html

Thus, we need to call the mapping function differently if you want to
convert `EL` to `GR` for getting the European countries than to pass
the proper country code.

This is done adding an extra parameter for saying if we want to include
such replacements, and only being executed when the check about European
country is rejected, on contrary than with Greece (EL), that should be
included as intracommunity.

The country code is now also obtained from the vat itself if the first
2 letter belongs to any country code.

TT38835