[FIX] l10n_es_aeat_sii_oca: Get VAT info from delivery address
Odoo determines the fiscal position, and thus, the taxes, through the
invoice delivery address (when filled).

But we are always using the commercial partner for getting the VAT info
(country code, VAT, etc), leading to inconsistencies.

With this commit, we amend it using when available such delivery address
as the source for computing the VAT info. It uses a little hack for
introspecting if the field is available in the model for not requiring
`sale` module as dependency or to create a glue module just for this.

TT38835