[FIX] l10n_es_pos: offline simplified number
In 2967122efbbe6445b239d07cb43e874c2647c1ba the behavior of the
simplified sequence was change to gather via rpc for each ticket
validation. For the case of offline the planned fallback was to set the
unique ID given by Odoo.

This change fallbacks to the former behavior for that case allowing to
obtain the proper sequence even offline.

The main drawback will be that concurrent users in the same session
could collide their sequence numbers if they happen to be offline.
Anyway, that would be easily fixed just by separating such cashiers in
separate configs.

As a solution to that, a different technique should be used (longpoll
for multi sessions or something like that)