[FIX] l10n_es_aeat: Avoid error on record initialization
On the v13 ORM, the related fields are not initialized on time before
doing the rest of the computations, including computed fields, so we
get an empty currency for doing the rounds, obtaining a crash.

For avoiding this, let's include a default value on the related field
that allows it to be initialized on time.

This doesn't happen on v14+.

TT40824