[FIX] l10n_es_vat_book: Fill correctly special taxes data
Depending on the order of the journal items, sometimes the special
taxes section on the regular taxes lines were not filled correctly.

With this commit, we always make sure of them being filled without
performance penalty.

And as a plus, 2 tax lines fields were converted to computed stored for
better code comprehension.

TT44033