[FIX] l10n_es_aeat: Consider financial_type="other" as regular operations
Steps to reproduce the problem:

- PoS installed.
- Have a session with only payment methods with outstanding accounts (no cash payments received).
- Close and post session.
- The resulting journal entry is of type = "Other".

Such entries, even if having taxes, won't be included in the AEAT 303 model,
as the type = "Other" is not considered till now. That's because it can't
be deduced if the operation is regular or not.

Let's assume that such operations are always regular (as 99.9% of the times will be),
for not having differences in the amounts of the report vs balances in the tax.

TT44503