[FIX] l10n_es_account_statement_import_n43: Avoid partner mismatch
If we have very short concepts, there's a chance that the stripped
text on the subsequent bank patterns matches with a partner, without the
concept being really a partner.

Example: "2301IMPUESTOS", taking "ESTOS" as name to match, which can
match with "GESTOS S.L.".

For avoiding this, we put a minimal string length of 8 chars for making
the match.

TT44038