[FIX] l10n_es_aeat_sii_oca: Avoid invoice copy with type change in tests
There's a new integrity check for avoiding inconsistent states. We were
using a trick on tests for quickly copying a customer invoice,
and converting the copy to vendor bill, but data got inconsistent doing
it.

For what we were testing, that was not a problem, but with the new
check, the operation is aborted.

Thus, we create the vendor bill from scratch, refactoring a method for
doing it easily.

We also take the occasion to improve a bit the number of invoice copies
done.