[FIX] l10n_es_facturae: Allow invoicing users to set Facturae
As it was, only accountant users can set Facturae check, as the field
was inside a group with such restriction. It doesn't matter that the
field itself has another less restricting group.

Thus, we put the initial check in the same group as the rest, and
move the visibility attributes to each field.

TT47097