[FIX] l10n_es_facturae_face: upgrade mocks
In
https://github.com/mvantellingen/python-zeep/commit/decbb761a78c4ea06fecdcfb24d90e248cd3a3dd
zeep moved the `ServiceProxy` class to a different file, and our test
mocks started to fail. This made those tests flaky, making them randomly
fail, as you can see in
https://github.com/OCA/l10n-spain/pull/3394#issuecomment-1908583904 and
some other comments below.

I'm now setting a lower bound in the zeep requirement and mocking the
correct path to avoid those problems in the future.

In any case, there are still more possible connection problems that I
cannot fix at the moment because... well, the server is down and I
cannot obtain the WSDL file. For those cases, we will just record the
failure and skip the test.

I took the chance to DRY all those mocks and upgrade some outdated
import patterns that are not relevant in Odoo 16 anymore.

These tests are now labeled with the correct tags, so they can be
completely skipped for those of us that don't care about external tests.