[FIX] l10n_es_payment_order_confirming_aef: Put processing date as today
Bankinter is actually using the processing date indicated in the file for their
planning and using the first line payment date causes that the files generated
don't work with that bank.

As harmless for the rest of the banks, let's use the date where the file is generated
for this data.

Closes #3425