[IMP] l10n_es_vat_book + l10n_es_aeat: Drill through on tax summary
Allow to drill through in the tax summary to list journal items that
leads to the base and tax amounts.

As we are on this, there's a tool method now in base module l10n_es_aeat
for opening this kind of view, and fixing that the journal items was
always open with move grouping by default.

TT52996