[FIX] l10n_es_aeat_mod303: Remove field 96
It was incorrectly mapped to "recargo de equivalencia" base amounts,
which is not correct, as that base amounts are already included in
regular operations, so it appears duplicated.

This field is for being filled in specific cases where the subject is
on the special fiscal regime "Recargo de equivalencia".

The tests have been changed, as now field 88 doesn't contain this field.

TT52996