[12.0][FIX] l10n_es_aeat_mod303: New BOE format 2025
In fact, the changes are very small for the implemented things, as
the changes on the REAGP, which are the biggest, are not implemented.
We just need to force to 0 the percentages for the ephemeral taxes.
They just can ignore that fields for not bothering people...

There's also a move for "Recargo equivalencia 0.5%" from 16-18 to
168-170 inadvertently.