[FIX] l10n_es_partner: display address with commercial name
Before this commit the commercial name would be display at the
end of the address. Something like this:

Paco Poco S.A.
Calle Mayor, 1
50001 Zaragoza
España (Industrias Paco)

Which is not the expected way. With this change, we can show it after the partner's name:

Paco Poco S.A (Industrias Paco)
Calle Mayor, 1
50001 Zaragoza
España

MT-8475