[FIX] l10n_es_vat_prorate: Avoid false prorate lines
When you publish a vendor bill with prorate taxes, and then reset it to
draft and change amounts, there's a chance that Odoo rewrites the
move lines switching the prorated expense and the tax, causing that
the new tax line, being the old expense one, have the flag `vat_prorate`
set, and thus not being included in AEAT reports.

As it seems we can't do anything for assuring the same move lines
processing order upstream, what we do is to force in all the entries
the False value, and then only putting to True when building the
VAT prorate expense line.

TT57089