[IMP] l10n_es_vat_book*: Add deductible amount + being compatible with l10n_es_vat_prorate
The tax fee amount is not always the same as the deductible amount when
your company is in the prorate regime.

AEAT splits both amounts in the corresponding columns, and we previously
were setting the same value on both.

After this commit, there's a separated field for the deductible amount,
and data is computed through the standard methods from `l10n_es_aeat`,
so when plugging `l10n_es_vat_prorate`, the deductible amount is
automatically changed by this module.

The refactoring needed for this to happen has been the minimum one for
minimizing the possible impacts, only affecting in the creation in
memory of one extra dictionary per vat book tax line, which is
acceptable. A deeper refactorization can rename fields of
"l10n.es.vat.book.line.tax" to match the returned values of the
standard methods for not having double dictionaries.

The new field is shown also on the UI, used for the XLS export, and
a migration script fill it with the same value as the amount for the
existing records.

A change of the map lines mapping has also been needed: tax_account_id
has been replaced by tax_excluded_account_id, reversing their usage:
now it excludes the account set instead of including, as if not, we
were filtering out the lines of the prorates. The renaming is on purpose
to warn people extending this module when they have used the previous
field, as updating this, they will have an error.

The non deductible VAT has been also adjusted according what we have
checked in the AEAT: the fee is reported, but the deductible amount is
0.