Display the FACE sending method in children opened directly
Partners childs facturae_sending_code is visible if we open the view from the parent partner (via child_ids) but not when we open the child partner directly. This PR correct this.

Since https://github.com/OCA/l10n-spain/pull/3372, in views without accounting information, there is a new view group with invoice information.

As far this module depends on l10n_es_facturae, it should inherit this behavior. The only change is to add facturae_sending_code to the same view group "group_facturae_wo_accounting" as the rest of the facturae information.