[FIX] l10n_es_intrastat_report: wrong country code for csv export
When export csv file is generated, calculated origin country code for
declaration lines is ignored, using default country code instead. This
error arises e.g. when a UK product is selected, saving at csv file GB
instead of XU. This fixes it.

For later versions of this addon, this is already fixed.