[FIX] l10n_es_vat_prorate: Fix module after special prorate update
After update module with special prorate (https://github.com/OCA/l10n-spain/pull/4287), the `prorate_id` field is not set in any of the previous journal entries, nor is the `with_vat_prorate` field set in the previous journal items.

This behavior makes some errors, for example, if you add a credit note of the invoice before the updating, the journal item will have `with_vat_prorate` field  as false because Odoo copy vals and the field is not recomputed.

With this migration script, `prorate_id` and `with_vat_prorate` are set.

MT-11679 @moduon