[FIX] l10n_es_partner: Enable searching by the commercial field in the search view
The field used in the search view is complete_name, but this field cannot be inherited because, in this line of code:
https://github.com/odoo/odoo/blob/5238a24752de83ef347755dc6c2d5f9c6f3e6772/odoo/addons/base/models/res_partner.py#L381

Odoo calls the _get_complete_name function without context.

When a user searches in the Contacts menu by the commercial name, no results are found.
With this commit, we inject an additional domain when a search for complete_name is performed, enabling results to be found by the commercial name.